GNUmakefile: Address complaint in file comments
This Makefile had unconditional checks for libbsd and ncurses/terminfo
libraries that result in failure even when not attempting to link mg,
and complained (rightly) that it was inappropriate to do so when trying
to make the "clean" or "uninstall" targets.  Fix that.

Make the target that links the executable depend on a phony target that
performs equivalent checks, and uses shell logic instead of GNU Make
features to test the Make macros that populate linker flags.

.PHONY is a POSIX 2024 feature.

Also update mention of "libncurses5-dev" to "libncurses6-dev".  ncurses
6.0 has been out since August 2015.