Trivially refactor (`map_element` -> `map_elt`)
Rename `map_element` struct to `map_elt`.  The code style is to use tabs
really heavily, for indentation as well as alignment of declarations,
which makes lengthy identifier names (for types and objects alike)
cumbersome.

Slightly recast comment annotating `remap()` since we needed to update
it for the name change anyway.